requirements.txt
feedparser>=6.0,<7.0
arxiv>=2.0,<3.0
requests>=2.28,<3.0
beautifulsoup4>=4.12,<5.0
jinja2>=3.1,<4.0
pyyaml>=6.0,<7.0
